feat(code): optimize docs based on review comment
