+++ packages/varlet-ui/src/code/example/index.vue
@@ -1,40 +1,42 @@
 <script setup>
-import { ref } from 'vue'
 import { t, use } from './locale'
 import { AppType, watchLang, onThemeChange } from '@varlet/cli/client'
 import { codeToHtml } from 'shiki'
+import hljs from 'highlight.js'
+import javascript from 'highlight.js/lib/languages/javascript'
 
-const jsCode = `function twoSum(nums, target) {
-  const map = new Map();
-  for (let i = 0; i < nums.length; i++) {
-      const theOther = target - nums[i];
-      if (map.has(theOther)) {
-          return [map.get(theOther), i];
-      }
-      map.set(nums[i], i);
-  }
-};`
+hljs.registerLanguage('javascript', javascript)
 
-const javaCode = `class Solution {
-  public int[] twoSum(int[] nums, int target) {
-    Map<Integer, Integer> map = new HashMap<>();
-    for (int i = 0; i < nums.length; i++) {
-      int theOther = target - nums[i];
-      if (map.containsKey(theOther)) {
-        return new int[] { map.get(theOther), i };
-      }
-      map.put(nums[i], i);
-    }
-    throw new IllegalArgumentException("No two sum solution");
-  }
-}`
-
-const language = ref('javascript')
-const theme = ref('monokai')
-
-function createHighlighter() {
+function createShikiHighlighter() {
   return {
     codeToHtml,
+  }
+}
+
+function createHljsHighlighter() {
+  const loadedThemes = new Set()
+
+  return {
+    codeToHtml: async (code, { lang, theme }) => {
+      if (!loadedThemes.has(loadedThemes)) {
+        const response = await fetch(
+          `https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.6.0/styles/${theme}.min.css`
+        )
+        const cssText = await response.text()
+        const modifiedCss = cssText.replace(/[^{.*?}]+(?=\{)|(?<=\})[^{}]+/g, (s) => `.${theme} ${s}`)
+        const style = document.createElement('style')
+        style.textContent = modifiedCss
+        document.head.appendChild(style)
+        loadedThemes.add(theme)
+      }
+
+      const highlightedCode = hljs.highlight(code, { language: lang }).value
+      return `
+        <div class="${theme}">
+          <pre class="hljs"><code>${highlightedCode}</code></pre>
+        </div>
+      `
+    },
   }
 }
 
@@ -43,19 +45,17 @@ onThemeChange()
 </script>
 
 <template>
-  <app-type>{{ t('language') }}</app-type>
-  <var-select :hint="false" v-model="language">
-    <var-option label="javascript" value="javascript" />
-    <var-option label="java" value="java" />
-  </var-select>
+  <app-type>{{ t('shiki') }}</app-type>
+  <var-highlighter-provider :highlighter="createShikiHighlighter()" theme="nord">
+    <var-code code="console.log('varlet')" language="javascript" />
+    <var-code code="console.log('varlet')" language="javascript" theme="monokai" />
+    <var-code code="console.log('varlet')" language="javascript" theme="one-dark-pro" />
+  </var-highlighter-provider>
 
-  <app-type>{{ t('theme') }}</app-type>
-  <var-select :hint="false" v-model="theme">
-    <var-option label="monokai" value="monokai" />
-    <var-option label="nord" value="nord" />
-  </var-select>
-
-  <var-highlighter-provider :highlighter="createHighlighter()">
-    <var-code :code="language === 'javascript' ? jsCode : javaCode" :language="language" :theme="theme" />
+  <app-type>{{ t('hljs') }}</app-type>
+  <var-highlighter-provider :highlighter="createHljsHighlighter()" theme="nord">
+    <var-code code="console.log('varlet')" language="javascript" />
+    <var-code code="console.log('varlet')" language="javascript" theme="monokai" />
+    <var-code code="console.log('varlet')" language="javascript" theme="atom-one-dark" />
   </var-highlighter-provider>
 </template>
