--- packages/varlet-ui/src/highlighter-provider/example/locale/en-US.ts
@@ -1,4 +1,0 @@
-export default {
-  shiki: 'Use Shiki',
-  hljs: 'Use highlight.js',
-}
