--- packages/varlet-ui/src/highlighter-provider/example/locale/zh-CN.ts
@@ -1,4 +1,0 @@
-export default {
-  shiki: '使用 Shiki',
-  hljs: '使用 highlight.js',
-}
